+++ ScriptEditor/FormCastsEditor.cs
@@ -613,7 +613,7 @@ namespace ScriptEditor
         }
 
         // Generic function for setting script field to specified value;
-        private void SetScriptFieldFromValue(float fieldvalue, string fieldname)
+        private void SetScriptFieldFromValue(double fieldvalue, string fieldname)
         {
             if (lstSpellLists.SelectedItems.Count > 0)
             {
@@ -639,8 +639,8 @@ namespace ScriptEditor
                 return;
 
             // Get the value from the textbox.
-            float fieldValue;
-            float.TryParse(ctrl.Text, out fieldValue);
+            double fieldValue;
+            double.TryParse(ctrl.Text, out fieldValue);
 
             // Set the field value.
             SetScriptFieldFromValue(fieldValue, fieldname);
